Permit the use of service name or protocol port.
While adding a port for an AV scanner, it didn't have a named port so we couldn't add it to the 'rhel8stig_white_list_services' list variable.  Adding the "regex_search()" filter automatically uses the correct module parameter based on the value of each item in the list variable.